Refactor presenter factory implementations and replace with `invoke` interface methods:
- Remove dependency-based initializations in factory classes.
- Simplify testing by introducing `Fake*Presenter` classes with `TODO` implementations.
- Update tests, imports, and configurations to align with interface-based contracts.
