+++ presenter/home/src/commonTest/kotlin/com/thomaskioko/tvmaniac/presenter/home/HomePresenterTest.kt
@@ -4,30 +4,11 @@ import app.cash.turbine.test
 import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.essenty.lifecycle.LifecycleRegistry
 import com.arkivanov.essenty.lifecycle.resume
-import com.thomaskioko.tvmaniac.core.base.model.AppCoroutineDispatchers
-import com.thomaskioko.tvmaniac.core.logger.fixture.FakeLogger
-import com.thomaskioko.tvmaniac.data.featuredshows.api.interactor.FeaturedShowsInteractor
-import com.thomaskioko.tvmaniac.data.featuredshows.testing.FakeFeaturedShowsRepository
-import com.thomaskioko.tvmaniac.data.popularshows.api.PopularShowsInteractor
-import com.thomaskioko.tvmaniac.data.popularshows.testing.FakePopularShowsRepository
-import com.thomaskioko.tvmaniac.data.topratedshows.testing.FakeTopRatedShowsRepository
-import com.thomaskioko.tvmaniac.data.trendingshows.testing.FakeTrendingShowsRepository
-import com.thomaskioko.tvmaniac.data.upcomingshows.api.UpcomingShowsInteractor
-import com.thomaskioko.tvmaniac.data.upcomingshows.testing.FakeUpcomingShowsRepository
-import com.thomaskioko.tvmaniac.datastore.testing.FakeDatastoreRepository
-import com.thomaskioko.tvmaniac.discover.api.TrendingShowsInteractor
-import com.thomaskioko.tvmaniac.discover.presenter.di.DefaultDiscoverPresenterFactory
-import com.thomaskioko.tvmaniac.discover.presenter.di.DiscoverPresenterFactory
-import com.thomaskioko.tvmaniac.domain.discover.DiscoverShowsInteractor
-import com.thomaskioko.tvmaniac.domain.genre.GenreShowsInteractor
-import com.thomaskioko.tvmaniac.genre.FakeGenreRepository
+import com.thomaskioko.tvmaniac.presenter.home.fakes.FakeDiscoverShowsPresenterFactory
 import com.thomaskioko.tvmaniac.presenter.home.fakes.FakeSearchPresenterFactory
 import com.thomaskioko.tvmaniac.presenter.home.fakes.FakeSettingsPresenterFactory
 import com.thomaskioko.tvmaniac.presenter.home.fakes.FakeWatchlistPresenterFactory
-import com.thomaskioko.tvmaniac.search.testing.FakeSearchRepository
-import com.thomaskioko.tvmaniac.topratedshows.data.api.TopRatedShowsInteractor
 import com.thomaskioko.tvmaniac.traktauth.testing.FakeTraktAuthManager
-import com.thomaskioko.tvmaniac.watchlist.testing.FakeWatchlistRepository
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.StandardTestDispatcher
@@ -40,21 +21,6 @@ class HomePresenterTest {
     private val lifecycle = LifecycleRegistry()
     private val testDispatcher = StandardTestDispatcher()
     private val traktAuthManager = FakeTraktAuthManager()
-    private val datastoreRepository = FakeDatastoreRepository()
-    private val featuredShowsRepository = FakeFeaturedShowsRepository()
-    private val trendingShowsRepository = FakeTrendingShowsRepository()
-    private val upcomingShowsRepository = FakeUpcomingShowsRepository()
-    private val topRatedShowsRepository = FakeTopRatedShowsRepository()
-    private val popularShowsRepository = FakePopularShowsRepository()
-    private val searchRepository = FakeSearchRepository()
-    private val genreRepository = FakeGenreRepository()
-    private val coroutineDispatcher = AppCoroutineDispatchers(
-        main = testDispatcher,
-        io = testDispatcher,
-        computation = testDispatcher,
-        databaseWrite = testDispatcher,
-        databaseRead = testDispatcher,
-    )
 
     private lateinit var presenter: HomePresenter
 
@@ -110,48 +76,10 @@ class HomePresenterTest {
 
     private fun buildHomePresenterFactory(): HomePresenter.Factory =
         DefaultHomePresenter.Factory(
-            discoverPresenterFactory = buildDiscoverPresenterFactory(),
+            discoverPresenterFactory = FakeDiscoverShowsPresenterFactory(),
             watchlistPresenterFactory = FakeWatchlistPresenterFactory(),
             searchPresenterFactory = FakeSearchPresenterFactory(),
             settingsPresenterFactory = FakeSettingsPresenterFactory(),
             traktAuthManager = traktAuthManager,
         )
-
-    private fun buildDiscoverPresenterFactory(): DiscoverPresenterFactory = DefaultDiscoverPresenterFactory(
-        discoverShowsInteractor = DiscoverShowsInteractor(
-            featuredShowsRepository = featuredShowsRepository,
-            topRatedShowsRepository = topRatedShowsRepository,
-            popularShowsRepository = popularShowsRepository,
-            trendingShowsRepository = trendingShowsRepository,
-            upcomingShowsRepository = upcomingShowsRepository,
-            genreRepository = genreRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        watchlistRepository = FakeWatchlistRepository(),
-        featuredShowsInteractor = FeaturedShowsInteractor(
-            featuredShowsRepository = featuredShowsRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        topRatedShowsInteractor = TopRatedShowsInteractor(
-            topRatedShowsRepository = topRatedShowsRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        popularShowsInteractor = PopularShowsInteractor(
-            popularShowsRepository = popularShowsRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        trendingShowsInteractor = TrendingShowsInteractor(
-            trendingShowsRepository = trendingShowsRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        upcomingShowsInteractor = UpcomingShowsInteractor(
-            upcomingShowsRepository = upcomingShowsRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        genreShowsInteractor = GenreShowsInteractor(
-            repository = genreRepository,
-            dispatchers = coroutineDispatcher,
-        ),
-        logger = FakeLogger(),
-    )
 }
