+++ presenter/home/src/commonTest/kotlin/com/thomaskioko/tvmaniac/presenter/home/fakes/FakeWatchlistPresenterFactory.kt
@@ -1,19 +1,23 @@
 package com.thomaskioko.tvmaniac.presenter.home.fakes
 
 import com.arkivanov.decompose.ComponentContext
+import com.thomaskioko.tvmaniac.watchlist.presenter.WatchlistAction
 import com.thomaskioko.tvmaniac.watchlist.presenter.WatchlistPresenter
-import com.thomaskioko.tvmaniac.watchlist.presenter.di.WatchlistPresenterFactory
-import com.thomaskioko.tvmaniac.watchlist.testing.FakeWatchlistRepository
+import com.thomaskioko.tvmaniac.watchlist.presenter.WatchlistState
+import kotlinx.coroutines.flow.StateFlow
 
-class FakeWatchlistPresenterFactory : WatchlistPresenterFactory {
-    val repository = FakeWatchlistRepository()
+class FakeWatchlistPresenterFactory : WatchlistPresenter.Factory {
 
-    override fun create(
+    override fun invoke(
         componentContext: ComponentContext,
         navigateToShowDetails: (showDetails: Long) -> Unit,
-    ): WatchlistPresenter = WatchlistPresenter(
-        componentContext = componentContext,
-        navigateToShowDetails = navigateToShowDetails,
-        repository = repository,
-    )
+    ): WatchlistPresenter = FakeWatchlistPresenter()
+}
+
+internal class FakeWatchlistPresenter : WatchlistPresenter {
+    override val state: StateFlow<WatchlistState>
+        get() = TODO("Not yet implemented")
+
+    override fun dispatch(action: WatchlistAction) {
+    }
 }
