+++ presenter/more-shows/src/commonTest/kotlin/com/thomaskioko/tvmaniac/moreshows/presentation/MoreShowsPresenterTest.kt
@@ -8,5 +8,5 @@ class MoreShowsPresenterTest {
     private val lifecycle = LifecycleRegistry()
     private val testDispatcher = StandardTestDispatcher()
 
-    private lateinit var presenter: MoreShowsPresenter
+    private lateinit var presenter: DefaultMoreShowsPresenter
 }
