--- presenter/search/src/commonMain/kotlin/com/thomaskioko/tvmaniac/search/presenter/di/SearchPresenterFactory.kt
@@ -1,41 +1,0 @@
-package com.thomaskioko.tvmaniac.search.presenter.di
-
-import com.arkivanov.decompose.ComponentContext
-import com.thomaskioko.tvmaniac.core.base.annotations.ActivityScope
-import com.thomaskioko.tvmaniac.genre.GenreRepository
-import com.thomaskioko.tvmaniac.search.api.SearchRepository
-import com.thomaskioko.tvmaniac.search.presenter.Mapper
-import com.thomaskioko.tvmaniac.search.presenter.SearchShowsPresenter
-import me.tatarka.inject.annotations.Inject
-import software.amazon.lastmile.kotlin.inject.anvil.ContributesBinding
-import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
-
-interface SearchPresenterFactory {
-    fun create(
-        componentContext: ComponentContext,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-        onNavigateToGenre: (id: Long) -> Unit,
-    ): SearchShowsPresenter
-}
-
-@Inject
-@SingleIn(ActivityScope::class)
-@ContributesBinding(ActivityScope::class, SearchPresenterFactory::class)
-class DefaultSearchPresenterFactory(
-    private val mapper: Mapper,
-    private val searchRepository: SearchRepository,
-    private val genreRepository: GenreRepository,
-) : SearchPresenterFactory {
-    override fun create(
-        componentContext: ComponentContext,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-        onNavigateToGenre: (id: Long) -> Unit,
-    ): SearchShowsPresenter = SearchShowsPresenter(
-        componentContext = componentContext,
-        onNavigateToShowDetails = onNavigateToShowDetails,
-        onNavigateToGenre = onNavigateToGenre,
-        mapper = mapper,
-        searchRepository = searchRepository,
-        genreRepository = genreRepository,
-    )
-}
