+++ presenter/search/src/commonTest/kotlin/com/thomaskioko/tvmaniac/presenter/search/SearchShowsPresenterTest.kt
@@ -7,6 +7,7 @@ import com.thomakioko.tvmaniac.util.testing.FakeFormatterUtil
 import com.thomaskioko.tvmaniac.genre.FakeGenreRepository
 import com.thomaskioko.tvmaniac.genre.ShowGenresEntity
 import com.thomaskioko.tvmaniac.search.presenter.ClearQuery
+import com.thomaskioko.tvmaniac.search.presenter.DefaultSearchShowsPresenter
 import com.thomaskioko.tvmaniac.search.presenter.EmptySearchResult
 import com.thomaskioko.tvmaniac.search.presenter.InitialSearchState
 import com.thomaskioko.tvmaniac.search.presenter.Mapper
@@ -387,7 +388,7 @@ class SearchShowsPresenterTest {
 
     private fun buildPresenter(
         lifecycle: LifecycleRegistry = LifecycleRegistry(),
-    ): SearchShowsPresenter = SearchShowsPresenter(
+    ): SearchShowsPresenter = DefaultSearchShowsPresenter(
         componentContext = DefaultComponentContext(lifecycle = lifecycle),
         onNavigateToShowDetails = {},
         onNavigateToGenre = {},
